+++ libs/agno/agno/memory/manager.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, Dict, List, Optional
 
 from agno.db.schemas.memory import MemoryRow
 from agno.memory.db.base import MemoryDb
-from agno.memory import UserMemory
+from agno.memory.schemas import UserMemory
 from agno.models.base import Model
 from agno.models.message import Message
 from agno.tools.function import Function
@@ -346,8 +346,6 @@ class MemoryManager:
             """
             from uuid import uuid4
 
-            from agno.memory.memory import UserMemory
-
             try:
                 last_updated = datetime.now()
                 memory_id = str(uuid4())
@@ -380,7 +378,6 @@ class MemoryManager:
             Returns:
                 str: A message indicating if the memory was updated successfully or not.
             """
-            from agno.memory.memory import UserMemory
 
             try:
                 last_updated = datetime.now()
