+++ libs/agno/agno/memory/schemas.py
@@ -1,0 +1,84 @@
+from dataclasses import dataclass, field
+from datetime import datetime
+from typing import Any, Dict, List, Optional, Union
+
+from pydantic import BaseModel, Field
+
+from agno.run.response import RunResponse
+from agno.run.team import TeamRunResponse
+
+
+class MemorySearchResponse(BaseModel):
+    """Model for Memory Search Response."""
+
+    memory_ids: List[str] = Field(
+        ..., description="The IDs of the memories that are most semantically similar to the query."
+    )
+
+
+@dataclass
+class UserMemory:
+    """Model for User Memories"""
+
+    memory: str
+    topics: Optional[List[str]] = None
+    input: Optional[str] = None
+    last_updated: Optional[datetime] = None
+    memory_id: Optional[str] = None
+
+    def to_dict(self) -> Dict[str, Any]:
+        _dict = {
+            "memory_id": self.memory_id,
+            "memory": self.memory,
+            "topics": self.topics,
+            "last_updated": self.last_updated.isoformat() if self.last_updated else None,
+            "input": self.input,
+        }
+        return {k: v for k, v in _dict.items() if v is not None}
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> "UserMemory":
+        last_updated = data.get("last_updated")
+        if last_updated:
+            data["last_updated"] = datetime.fromisoformat(last_updated)
+        return cls(**data)
+
+
+@dataclass
+class TeamMemberInteraction:
+    member_name: str
+    task: str
+    response: Union[RunResponse, TeamRunResponse]
+
+    def to_dict(self) -> Dict[str, Any]:
+        return {
+            "member_name": self.member_name,
+            "task": self.task,
+            "response": self.response.to_dict(),
+        }
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> "TeamMemberInteraction":
+        return cls(member_name=data["member_name"], task=data["task"], response=RunResponse.from_dict(data["response"]))
+
+
+@dataclass
+class TeamContext:
+    # List of team member interaction, represented as a request and a response
+    member_interactions: List[TeamMemberInteraction] = field(default_factory=list)
+    text: Optional[str] = None
+
+    def to_dict(self) -> Dict[str, Any]:
+        return {
+            "member_interactions": [interaction.to_dict() for interaction in self.member_interactions],
+            "text": self.text,
+        }
+
+    @classmethod
+    def from_dict(cls, data: Dict[str, Any]) -> "TeamContext":
+        return cls(
+            member_interactions=[
+                TeamMemberInteraction.from_dict(interaction) for interaction in data["member_interactions"]
+            ],
+            text=data["text"],
+        )
