+++ libs/agno/agno/team/team.py
@@ -33,7 +33,6 @@ from agno.exceptions import ModelProviderError, RunCancelledException
 from agno.knowledge.agent import AgentKnowledge
 from agno.media import Audio, AudioArtifact, AudioResponse, File, Image, ImageArtifact, Video, VideoArtifact
 from agno.memory.memory import Memory
-from agno.session.summarizer import SessionSummary
 from agno.models.base import Model
 from agno.models.message import Citations, Message, MessageReferences
 from agno.models.response import ModelResponse, ModelResponseEvent, ToolExecution
@@ -44,6 +43,7 @@ from agno.run.response import RunResponse
 from agno.run.team import TeamRunEvent, TeamRunResponse, TeamRunResponseEvent, ToolCallCompletedEvent
 from agno.session import TeamSession
 from agno.session.metrics import SessionMetrics
+from agno.session.summarizer import SessionSummary
 from agno.tools.function import Function
 from agno.tools.toolkit import Toolkit
 from agno.utils.events import (
@@ -6268,7 +6268,7 @@ class Team:
                 except Exception as e:
                     log_warning(f"Failed to load runs from memory: {e}")
             if "team_context" in session.memory:
-                from agno.memory.memory import TeamContext
+                from agno.memory.schemas import TeamContext
 
                 try:
                     self.memory.team_context = {
@@ -6281,7 +6281,7 @@ class Team:
                 if self.memory.memories is not None:
                     pass
                 else:
-                    from agno.memory.memory import UserMemory as UserMemoryV2
+                    from agno.memory.schemas import UserMemory as UserMemoryV2
 
                     try:
                         self.memory.memories = {
